cortex_agent_a2a/requirements.txt
fastapi>=0.110.0
uvicorn>=0.30.0
cryptography>=41.0.0
PyJWT>=2.8.0
httpx>=0.27.0
python-dotenv>=1.0.0
requests>=2.31.0
snowflake-connector-python>=3.6.0
